moved prime_factorization to factors module to avoid circular depenancy issues
