+++ valkyrie_util/primes.py
@@ -1,7 +1,6 @@
 '''utility generators and functions related to primes'''
 
-from typing import Dict, Generator, List
-from valkyrie_util.factors import combine_factorizations
+from typing import Generator, List
 
 
 class _smart_sieve:
@@ -59,40 +58,6 @@ class _smart_sieve:
         return len(_smart_sieve.__sieve)
 
 
-class _smart_factorization:
-    '''internal memory of factorizations,
-
-    ensures no recalculations of factorizations'''
-
-    known_factorizations: Dict[int, Dict[int, int]] = {1: {}}
-
-    @staticmethod
-    def get_prime_factors(n: int, last_prime_tested: int = 1) -> Dict[int, int]:
-        '''recursively calculates factorizations'''
-        if n in _smart_factorization.known_factorizations:
-            return _smart_factorization.known_factorizations[n].copy()
-
-        current_divisor = nth_prime(last_prime_tested)
-
-        # add the prime to factorization if not already in
-        if current_divisor not in _smart_factorization.known_factorizations:
-            _smart_factorization.known_factorizations[current_divisor] = {current_divisor: 1}
-
-        # can stop and conclude that n is prime if current_divisor ** 2 > n
-        if current_divisor * current_divisor > n:
-            _smart_factorization.known_factorizations[n] = {n: 1}
-            return {n: 1}
-
-        if n % current_divisor == 0:
-            quotient = n // current_divisor
-            quotient_fact = _smart_factorization.get_prime_factors(quotient, last_prime_tested)
-            fact = combine_factorizations([{current_divisor: 1}, quotient_fact])
-            _smart_factorization.known_factorizations[n] = fact
-            return fact
-        else:
-            return _smart_factorization.get_prime_factors(n, last_prime_tested + 1)
-
-
 def prime_iterator() -> Generator[int, None, None]:
     '''iterates through all primes starting at 2'''
     n = 1
@@ -122,12 +87,3 @@ def nth_prime(n: int) -> int:
     if n < 1:
         raise ValueError
     return _smart_sieve.nth_prime(n)
-
-
-def prime_factorization(n: int) -> Dict[int, int]:
-    '''returns the prime factorization of n where the n equals the product of all keys ** value'''
-    if n < 1:
-        raise ValueError
-    elif n == 1:
-        return {}
-    return _smart_factorization.get_prime_factors(n)
